+++ src/SDKs/DataFactory/Management.DataFactory/changelog.md
@@ -9,3 +9,4 @@
 
 ### Feature Additions
   * Initial public release of the Azure Data Factory V2 .NET SDK.
+  * Add a new AzureMySql connector for Azure MySQL data source.